src/app_text2sql/pyproject.toml
[tool.poetry]
name = "app-text2sql"
version = "0.1.0"
description = ""
authors = ["Simon Jakesch <simonjakesch@microsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.36"
llama-index-core = "^0.11.22"
llama-index-llms-azure-openai = "^0.2.2"
asyncpg = "^0.30.0"
asyncio = "^3.4.3"
llama-index-embeddings-azure-openai = "^0.2.5"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
